don't pull in all of calypso-ui for screen-reader-text wrapper
This is a terrible hacky work-around for an issue in @wordpress/components. @wordpress/components relies on @wordpress/priority-queue, which tries to use things on window directly. This breaks when it gets pulled into our build/bundle by the server webpack build, the build cannot start. By adding text-control to index, this transitively pulls in @wordpress/components, which pulls in priority-queue, which breaks. Work around it by going after screen-reader-text directly.

This is not a long term solution. Ideally wordpress components would work from within a node process.